refactor(sidebar): SidebarPane の責務を composable とコンポーネントに分離する
- Extract useSidebarData: データ取得・状態管理・computed・イベントリスナー
- Extract useWorktreeActions: worktree CRUD・選択・切り替え（独自 isCreating で re-entry guard）
- Extract useTodoActions: Todo CRUD・インライン編集
- Extract useDialogs: 確認/通知ダイアログの状態管理
- Extract useCtrlBadge: Ctrl キー押下検知
- Extract SidebarMenu.vue: ⋮ ポップオーバーメニュー（expose で openMenu を公開）
- Extract VoicevoxPanel.vue: VOICEVOX 操作パネル
- Add re-entry guard to submitNewWorktree and createWorktree
