+++ apps/renderer/src/features/sidebar/features/todo/index.ts
@@ -1,2 +1,3 @@
 export { default as TodoEditor } from "./TodoEditor.vue";
 export { default as TodoList } from "./TodoList.vue";
+export { useTodoActions } from "./useTodoActions";
