+++ apps/renderer/src/features/sidebar/features/worktree/index.ts
@@ -1,3 +1,4 @@
 export { default as BranchList } from "./BranchList.vue";
 export { default as RootWorktree } from "./RootWorktree.vue";
 export { default as WorktreeList } from "./WorktreeList.vue";
+export { useWorktreeActions } from "./useWorktreeActions";
